requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
numpy==1.26.4
openai==1.30.5
opencv-python==4.9.0.80
pydantic==2.7.2
pydantic_core==2.18.3
python-dotenv==1.0.1
redis==5.0.4
requests==2.32.3
sniffio==1.3.1
spotipy==2.24.0
tqdm==4.66.4
typing_extensions==4.12.1
urllib3==2.2.1
